Funcionando

+++ MovilTours/app/src/main/java/com/example/moviltours/Tour.kt
@@ -2,8 +2,8 @@ package com.example.moviltours
 
 data class Tour(
     val name: String,
-    val latitude: Float,
-    val longitude: Float,
+    val latitude: Double,
+    val longitude: Double,
     val price: String,
     val photo: String
 ) : java.io.Serializable
